feat(fullstack): Combine delete event files and few minor changes
