+++ app/server/src/features/organizations/methods.ts
@@ -5,7 +5,6 @@ import {
     DeleteOrganization,
     CreateMember,
     DeleteMember,
-    DeleteOrganization,
 } from '@local/graphql-types';
 import { isMemberOfOrg } from '@local/features/permissions';
 import { register } from '@local/features/accounts/methods';
